requirements.txt
streamlit==1.44.1
opencv-python==4.6.0.66
torch>=1.12.0
torchvision==0.22.0
ultralytics==8.3.115
filterpy==1.4.5
motmetrics==1.0.1  # For MOT (Multiple Object Tracking) metrics
numpy>=1.26.0,<2.2.0 
scipy==1.10.1  # For scientific computations
matplotlib==3.7.1  # For plotting and visualizations
pyyaml==6.0  # For YAML parsing (used for configuration files)
tqdm==4.65.0  # For progress bars during processing
pandas==2.0.0  # For handling dataframes
deep-sort-realtime==1.3.2  # For the Deep SORT tracking algorithm
bytetracker

